[front/components/assistant/conversation] - refactor: improve scroll behavior and clean up imports
 - Implement auto-scrolling to the bottom when a new message arrives using the CONVERSATION_PARENT_SCROLL_DIV_ID
 - Extract IntersectionObserver logic to a custom hook for observing the last message group
 - Clean up and consolidate import statements from "@dust-tt/types" for better readability

[front/hooks] - feature: add custom hook for observing last message group

 - Create a useLastMessageGroupObserver hook to manage the visibility of the last message group efficiently
 - Decouple the message observation logic from the ConversationViewer component to enhance modularity
