+++ front/components/assistant/conversation/AgentMessage.tsx
@@ -27,8 +27,11 @@ import type {
   WebsearchResultType,
   WorkspaceType,
 } from "@dust-tt/types";
-import type { RetrievalDocumentType } from "@dust-tt/types";
-import type { AgentMessageType, MessageReactionType } from "@dust-tt/types";
+import type {
+  AgentMessageType,
+  MessageReactionType,
+  RetrievalDocumentType,
+} from "@dust-tt/types";
 import {
   assertNever,
   isRetrievalActionType,
